pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.31",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "groq>=0.13.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-groq>=0.2.0",
    "httpx>=0.27.0",
    "pydantic[email]>=2.12.5",
    "python-multipart>=0.0.21",
    "langchain-community>=0.4.1",
    "tavily-python>=0.7.17",
]
